The first CI run this repo ever had failed on all three platforms. This is what
it took to answer PLAN section 486's milestone.

Four things MSVC never said: GCC's -Wextra on an unnamed enum constant sharing a
conditional with a uint16_t, Clang's -Wdouble-promotion mid-accumulation, a
<cmath> that only libstdc++ wants spelled out, and -Wnull-dereference on a decay
curve that really can be empty. macOS had never compiled at all, for want of
pthread_getcpuclockid; the thread registry carries a Mach port there now.

The budgets of PLAN section 9 describe the hardware the mod targets, so on
runners that are half that speed and share their cores they are measured,
printed and not gated -- VSA_NO_PERF_BUDGETS, honoured by the native tests and
by SceneLab, whose load expectation swung between 46 % and 137 % across two runs
of the same commit. The sanitizer job found nothing of ours: its leaks are
libvorbis's encoder and libphonon, now suppressed by a file ctest points at, and
third-party keeps ASan but loses UB checks it cannot act on.

Two answers the plan asked CI for. Steam Audio 4.8.1 has no Embree on Apple
Silicon (section 216), and IPL_SCENETYPE_DEFAULT there bakes the 64^3 pathing
region in 323 s against 0.57 s on x64. And its built-in HRTF loses nearly all of
the rear roll-off on arm64 -- the same KEMAR SOFA file renders the rear within
0.001 dB on both, so it is the HRTF and not the convolution;
docs/investigations/steam-audio-arm64-hrtf.md has what was ruled out.

+++ native/src/api.cpp
@@ -688,7 +688,7 @@ VSA_API vsa_result VSA_CALL vsa_engine_get_sources(vsa_engine* engine, vsa_sourc
             const vsa::world::SourceDebug& s = sources[i];
             vsa_source_debug d{};
             d.struct_size = sizeof d;
-            d.flags = s.escaped ? VSA_SOURCE_ESCAPED : 0u;
+            d.flags = s.escaped ? static_cast<uint32_t>(VSA_SOURCE_ESCAPED) : 0u;
             d.voice = s.voice;
             std::copy_n(s.position, 3, d.position);
             std::copy_n(s.simulated_position, 3, d.simulated_position);
